requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
alembic==1.8.1
Babel==2.10.3
chardet
click==8.1.3
DateTimeRange==1.2.0
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Table==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mbstrdecoder==1.1.0
paho-mqtt==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
typepy==1.3.0
Werkzeug==2.1.2